+++ changelog.md
@@ -21,6 +21,7 @@ Development
 + Removed NetworkX requirement.
 + Two functions are moved to core kernel. ("edges_view" and "graph2vpoints")
 + Add "--platform" argument for Qt plugins.
++ Optimization of joint type of "VPoint" class from core kernel.
 + Update SHELL variable of Makefile on Windows platform.
 + Environment of Visual Studio 2017 version 15.8 supported. (`_MSC_VER == 1915`)